GIT PUSH

+++ The Seekers Way 0.2/State.h
@@ -57,6 +57,6 @@ private:
 	bool mEsc;
 
 	void changeState();
-	int mMusicCount;
+	float mMusicCount;
 };
 #endif
